Add a more organized output format for encoding deferred imports.
Deferred imports have a library as a parent, and reference a list
of output units which are needed when the import is resolved.

Currently this information is not populated by dart2js, and I've
just manually filled what I expect the output to look like into
the test info.json file.