Make compatible with null-safe args package.
The return type of usageException() is Never in the migrated package.
Externally, dart2js_info does not need to worry about this because its
constraint on args keeps it on the unmigrated version.

But to roll the migrated args into the Dart SDK and google3, it needs to
be compatible with dart2js_info. This tiny change here makes the
usageException() override agnostic and work with both the unmigrated and
migrated versions of args. I tested the latter locally with a dependency
override.